refactor: move buildCsv method into service and format code
